refactor: extract DepartureRowContainer
